requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
dnspython==2.4.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

